feat: add new implementation crud in some methods in userservice and 100% test coverage improvement in Jacoco

+++ ccrHospitalManagement/src/main/java/com/example/ccrHospitalManagement/service/UserServiceImpl.java
@@ -1,18 +1,19 @@
 package com.example.ccrHospitalManagement.service;
 
-import com.example.ccrHospitalManagement.model.Role;
-import com.example.ccrHospitalManagement.model.User;
-import com.example.ccrHospitalManagement.model.UserRole;
-import com.example.ccrHospitalManagement.model.UserRoleId;
-import com.example.ccrHospitalManagement.repository.RoleRepository;
-import com.example.ccrHospitalManagement.repository.UserRepository;
-import com.example.ccrHospitalManagement.repository.UserRoleRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
+
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import com.example.ccrHospitalManagement.model.Role;
+import com.example.ccrHospitalManagement.model.User;
+import com.example.ccrHospitalManagement.repository.RoleRepository;
+import com.example.ccrHospitalManagement.repository.UserRepository;
+import com.example.ccrHospitalManagement.repository.UserRoleRepository;
+
+import lombok.RequiredArgsConstructor;
 
 @Service
 @RequiredArgsConstructor
@@ -67,7 +68,13 @@ public class UserServiceImpl implements UserService {
     @Override
     @Transactional
     public void deleteUser(String id) {
-
-        userRepository.deleteById(id);
+        Optional<User> user = userRepository.findById(id); // Buscar al usuario primero
+        if (user.isPresent()) {
+            userRepository.deleteById(id); // Si existe, lo eliminamos
+        } else {
+            // Si no existe, lanzar una excepción o simplemente no hacer nada
+            throw new IllegalArgumentException("Usuario no encontrado");
+        }
     }
+
 }
